IGNITE-24573 Move WriteIntentSwitchCommandHandler to the appropriate package
